feat: support the new package administration interface
You can opt into using the new flow by providing the `--useNewPackageAdministrationInterface`
command line flag or setting the `FVTT_USE_NEW_PACKAGE_ADMINISTRATION_INTERFACE` environment
variable to `true`. When using the new flow, the package ID needs to be the `id` from the manifest
instead of the numeric ID, which was needed for the legacy flow.